[versioned] Deal with packages that don't build on system
