+++ internal/impl/packages.go
@@ -150,12 +150,12 @@ const (
 func (d *Devbox) ensurePackagesAreInstalled(ctx context.Context, mode installMode) error {
 	defer trace.StartRegion(ctx, "ensurePackages").End()
 
-	lock, err := lock.Local(d)
+	localLock, err := lock.Local(d)
 	if err != nil {
 		return err
 	}
 
-	upToDate, err := lock.IsUpToDate()
+	upToDate, err := localLock.IsUpToDate()
 	if err != nil {
 		return err
 	}
@@ -183,7 +183,12 @@ func (d *Devbox) ensurePackagesAreInstalled(ctx context.Context, mode installMod
 		return err
 	}
 
-	return lock.Update()
+	if err = localLock.Update(); err != nil {
+		return err
+	}
+
+	// Update lockfile to ensure any newly resolved packages are saved to disk.
+	return d.lockfile.Save()
 }
 
 func (d *Devbox) profilePath() (string, error) {
